Fix email sharing, supabase admin, and Sentry logging (#22)
# Description

Added the `@supabase/supabase-js` library to create an admin instance to
bypass RLS policies when needed. This is because the Supabase NextJS
auth helpers use the JWT in the cookies as precedence over the service
role key. With the actual handling with Supabase being done in the
server, it is fine to use.

# Details

- Updated the logging for Sentry to include the `item_id` and map the
transactions to only include information to identify them for debugging
- Updated the CCF Records trends component to have larger font for the
headings
- Update the error message response for transactions sync to be generic
so it isn't sending sensitive information to the client
- Changed the `insert()` to `upsert()` for the transactions sync for
added transactions
- Update `global-env.d.ts` to include `CircleCI`