Plaid new accounts update mode (#116)
# Description

Added the case for the webhook when `NEW_ACCOUNTS_AVAILABLE` is
received. A new column `new_accounts` is added to the `plaid` table so
that when the user selects that institution, it triggers Plaid update
mode so the user can add the account.

## Details

- Updated `plaid` table with the `new_accounts` column
- `createLinkToken()` creates a Plaid `link_token` and can be given an
access token to create a link token that will be used for update mode
- Updated the `/api/plaid/link-token/create` endpoint to be able to
create the link token for update mode
- Updated fetches of data of institutions for the new `new_accounts`
property
- Renamed the original `create-link-token.ts` util file to
`get-link-token.ts` since it just fetches the token from the backend
- Update `use-plaid.tsx` so that it handles the account update from
Plaid
- Because I'm unsure of how the process flows, it doesn't actually make
changes except to set `new_accounts` to `false` and log the data.