Account Balances feature and fixes and improvements (#120)
# Description

Added the "Account Balances" feature under the "Finance Planning"
section where users can create accounts and track their balances with
the deposits. This is then used in the CCF Visualizer to display their
account balances over time.

## Details

- Update `Dialog` components to use `DialogClose`, where applicable
- Update the institutions/accounts to use `initcap()` in the hooks to
store the names formatted
- Updated `initcap.ts` so that it by default is strict, where it only
capitalizes the fist character of each word. If `strict` is set to
`false`, it will allow other characters to be capitalized as well.
- Fixed the institution `Command` combo-box so that it only renders the
list and the separator of items if there is a list and has a length.
- Update the `plaid` and `plaid_accounts` tables to lowercase the `name`
when checking to ensure there are no duplicates, regardless of casing.
- Added the name check into the function for the update trigger for
`plaid`
- Checks if the `user_id` or `item_id` matches first, before checking
the name for optimalisation
- Update Plaid `atoms.ts` so that the selected institution is updated on
any update
- Update `DateRangePicker` component to have custom button text and set
default text to `Select Dates`
- Update uses where it is used for filtering to have `"Filter Dates"`
text
- Fix clearing user on logout
	- Update `UserProvider` to clear data on unmount
	- Update `userAtom` with a `clearUserAtom` setter
	- Update `useLogout` so that is clears the `userAtom`
- Fix WaaInfo date filter state to be able to reset it
- Update `react-query.ts` constants
	- Added comments for what they're used for
	- Made the values uppercase
- Update CCF to use constant for react-query data
- Update the CCF Visualizer
- Calculation function waits one second when calculating to simulate
loading/processing
	- Using the `VISUALIZER_WAA_KEY` constant
	- Added the "Accounts Balances"