Plaid add accounts (#123)
# Description

Properly implemented the account update mode for Plaid. When the
institution selected has new accounts available, set from the webhook
and the `NEW_ACCOUNTS_AVAILABLE` code, it will fetch a new `link_token`
that uses a modified configuration to allow the user to select accounts.

This will then send the metadata received on successful re-link with
Plaid, to the backend, which will only add new accounts. The user must
have the new account selected to ensure all transaction data is
retrieved. Otherwise, an error will be thrown if there is no newly added
account when comparing the list of accounts and their IDs and, the ones
already in the database for the institution.

## Details
- `add_plaid_accounts` SQL function to add the new accounts and update
the `new_accounts` column on the institution after successful insertion
of the new account
  - Added new SQL custom type `add_plaid_account_type`
- After the new account(s) have been updated, it will re-fetch the
accounts data via react-query